chore(Cache): Make the XPC6 PR work with the renamed types
